requirements.txt
prompt_toolkit
rel
requests
websocket-client
